+++ changelog.md
@@ -3,3 +3,4 @@
 ##0.0.1-SNAPSHOT
 
 1. Added payment.txt file to load data in db
+2. Added payment object, added step to load dummy data in DB